+++ src/config.rs
@@ -9,7 +9,7 @@
 #![deny(missing_docs)]
 
 use std;
-use std::collections::{BTreeMap, BTreeSet};
+use std::collections::{BTreeMap, BTreeSet, HashMap};
 use std::fmt;
 use std::hash::Hash;
 use std::path::{Path, PathBuf};
@@ -21,6 +21,7 @@ use indexmap::IndexMap;
 use semver;
 use serde::de::{Deserialize, Deserializer};
 use serde::ser::{Serialize, Serializer};
+use serde_yaml::Value;
 use subst;
 
 use crate::error::*;
@@ -83,6 +84,8 @@ pub struct Package {
     /// A list of package authors. Each author should be of the form `John Doe
     /// <john@doe.com>`.
     pub authors: Option<Vec<String>>,
+    #[serde(flatten)]
+    extra: HashMap<String, Value>,
 }
 
 /// A dependency.
@@ -218,7 +221,7 @@ pub trait Validate {
     /// The error type produced by validation.
     type Error;
     /// Validate self and convert into the non-partial version.
-    fn validate(self) -> std::result::Result<Self::Output, Self::Error>;
+    fn validate(self, package_name: &str) -> std::result::Result<Self::Output, Self::Error>;
 }
 
 // Implement `Validate` for hash maps of validatable values.
@@ -229,9 +232,9 @@ where
 {
     type Output = IndexMap<K, V::Output>;
     type Error = (K, Error);
-    fn validate(self) -> std::result::Result<Self::Output, Self::Error> {
+    fn validate(self, package_name: &str) -> std::result::Result<Self::Output, Self::Error> {
         self.into_iter()
-            .map(|(k, v)| match v.validate() {
+            .map(|(k, v)| match v.validate(package_name) {
                 Ok(v) => Ok((k, v)),
                 Err(e) => Err((k, e)),
             })
@@ -245,9 +248,9 @@ where
 {
     type Output = Vec<V::Output>;
     type Error = Error;
-    fn validate(self) -> std::result::Result<Self::Output, Self::Error> {
+    fn validate(self, package_name: &str) -> std::result::Result<Self::Output, Self::Error> {
         self.into_iter()
-            .map(|v| match v.validate() {
+            .map(|v| match v.validate(package_name) {
                 Ok(v) => Ok(v),
                 Err(e) => Err(e),
             })
@@ -262,8 +265,8 @@ where
 {
     type Output = T::Output;
     type Error = T::Error;
-    fn validate(self) -> std::result::Result<T::Output, T::Error> {
-        self.0.validate()
+    fn validate(self, package_name: &str) -> std::result::Result<T::Output, T::Error> {
+        self.0.validate(package_name)
     }
 }
 
@@ -274,8 +277,8 @@ where
 {
     type Output = T::Output;
     type Error = T::Error;
-    fn validate(self) -> std::result::Result<T::Output, T::Error> {
-        self.0.validate()
+    fn validate(self, package_name: &str) -> std::result::Result<T::Output, T::Error> {
+        self.0.validate(package_name)
     }
 }
 
@@ -320,15 +323,18 @@ pub struct PartialManifest {
     pub workspace: Option<PartialWorkspace>,
     /// External Import dependencies
     pub vendor_package: Option<Vec<PartialVendorPackage>>,
+    /// Unknown extra fields
+    #[serde(flatten)]
+    extra: HashMap<String, Value>,
 }
 
 impl PartialManifest {
     /// Before fully cloning a git repo locally a manifest is read without the source files present. Since validate by default now
     /// checks to make sure source files actually exist this will error out when verifying a dependency manifest. To get around this,
     /// the initial check of a dependency manifest that is done before cloning the dependency will ignore whether source files exist
-    pub fn validate_ignore_sources(mut self) -> Result<Manifest> {
+    pub fn validate_ignore_sources(mut self, package_name: &str) -> Result<Manifest> {
         self.sources = Some(SeqOrStruct::new(PartialSources::new_empty()));
-        self.validate()
+        self.validate(package_name)
     }
 }
 
@@ -354,6 +360,7 @@ impl PrefixPaths for PartialManifest {
             frozen: self.frozen,
             workspace: self.workspace.prefix_paths(prefix)?,
             vendor_package: self.vendor_package.prefix_paths(prefix)?,
+            extra: self.extra,
         })
     }
 }
@@ -361,10 +368,17 @@ impl PrefixPaths for PartialManifest {
 impl Validate for PartialManifest {
     type Output = Manifest;
     type Error = Error;
-    fn validate(self) -> Result<Manifest> {
+    fn validate(self, _package_name: &str) -> Result<Manifest> {
         let pkg = match self.package {
             Some(mut p) => {
                 p.name = p.name.to_lowercase();
+                p.extra.iter().for_each(|(k, _)| {
+                    warnln!(
+                        "Ignoring unknown field `{}` in manifest package for {}.",
+                        k,
+                        p.name
+                    );
+                });
                 p
             }
             None => return Err(Error::new("Missing package information.")),
@@ -374,7 +388,7 @@ impl Validate for PartialManifest {
                 .into_iter()
                 .map(|(k, v)| (k.to_lowercase(), v))
                 .collect::<IndexMap<_, _>>()
-                .validate()
+                .validate(&pkg.name)
                 .map_err(|(key, cause)| {
                     Error::chain(
                         format!("In dependency `{}` of package `{}`:", key, pkg.name),
@@ -384,7 +398,7 @@ impl Validate for PartialManifest {
             None => IndexMap::new(),
         };
         let srcs = match self.sources {
-            Some(s) => Some(s.validate().map_err(|cause| {
+            Some(s) => Some(s.validate(&pkg.name).map_err(|cause| {
                 Error::chain(format!("In source list of package `{}`:", pkg.name), cause)
             })?),
             None => None,
@@ -400,16 +414,23 @@ impl Validate for PartialManifest {
         let frozen = self.frozen.unwrap_or(false);
         let workspace = match self.workspace {
             Some(w) => w
-                .validate()
+                .validate(&pkg.name)
                 .map_err(|cause| Error::chain("In workspace configuration:", cause))?,
             None => Workspace::default(),
         };
         let vendor_package = match self.vendor_package {
             Some(vend) => vend
-                .validate()
+                .validate(&pkg.name)
                 .map_err(|cause| Error::chain("Unable to parse vendor_package", cause))?,
             None => Vec::new(),
         };
+        self.extra.iter().for_each(|(k, _)| {
+            warnln!(
+                "Ignoring unknown field `{}` in manifest for {}.",
+                k,
+                pkg.name
+            );
+        });
         Ok(Manifest {
             package: pkg,
             dependencies: deps,
@@ -449,6 +470,9 @@ pub struct PartialDependency {
     /// The version requirement of the package. This will be parsed into a
     /// semantic versioning requirement.
     version: Option<String>,
+    /// Unknown extra fields
+    #[serde(flatten)]
+    extra: HashMap<String, Value>,
 }
 
 impl FromStr for PartialDependency {
@@ -459,6 +483,7 @@ impl FromStr for PartialDependency {
             git: None,
             rev: None,
             version: Some(s.into()),
+            extra: HashMap::new(),
         })
     }
 }
@@ -475,7 +500,7 @@ impl PrefixPaths for PartialDependency {
 impl Validate for PartialDependency {
     type Output = Dependency;
     type Error = Error;
-    fn validate(self) -> Result<Dependency> {
+    fn validate(self, package_name: &str) -> Result<Dependency> {
         let version = match self.version {
             Some(v) => Some(semver::VersionReq::parse(&v).map_err(|cause| {
                 Error::chain(
@@ -490,6 +515,13 @@ impl Validate for PartialDependency {
                 "A dependency cannot specify `version` and `rev` at the same time.",
             ));
         }
+        self.extra.iter().for_each(|(k, _)| {
+            warnln!(
+                "Ignoring unknown field `{}` in a dependency in manifest for {}.",
+                k,
+                package_name
+            );
+        });
         if let Some(path) = self.path {
             if let Some(list) = string_list(
                 self.git
@@ -538,6 +570,9 @@ pub struct PartialSources {
     pub defines: Option<IndexMap<String, Option<String>>>,
     /// The source file paths.
     pub files: Vec<PartialSourceFile>,
+    /// Unknown extra fields
+    #[serde(flatten)]
+    extra: HashMap<String, Value>,
 }
 
 impl PartialSources {
@@ -549,6 +584,7 @@ impl PartialSources {
             include_dirs: None,
             defines: None,
             files: Vec::new(),
+            extra: HashMap::new(),
         }
     }
 }
@@ -560,6 +596,7 @@ impl PrefixPaths for PartialSources {
             include_dirs: self.include_dirs.prefix_paths(prefix)?,
             defines: self.defines,
             files: self.files.prefix_paths(prefix)?,
+            extra: self.extra,
         })
     }
 }
@@ -571,6 +608,7 @@ impl From<Vec<PartialSourceFile>> for PartialSources {
             include_dirs: None,
             defines: None,
             files: v,
+            extra: HashMap::new(),
         }
     }
 }
@@ -578,7 +616,7 @@ impl From<Vec<PartialSourceFile>> for PartialSources {
 impl Validate for PartialSources {
     type Output = Sources;
     type Error = Error;
-    fn validate(self) -> Result<Sources> {
+    fn validate(self, package_name: &str) -> Result<Sources> {
         let post_glob_files: Vec<PartialSourceFile> = self
             .files
             .into_iter()
@@ -604,12 +642,29 @@ impl Validate for PartialSources {
             .map(|path| env_path_from_string(path.to_string()))
             .collect();
         let defines = self.defines.unwrap_or_default();
-        let files: Result<Vec<_>> = post_glob_files.into_iter().map(|f| f.validate()).collect();
+        let files: Result<Vec<_>> = post_glob_files
+            .into_iter()
+            .map(|f| f.validate(package_name))
+            .collect();
+        let files = files?;
+        if files.is_empty() {
+            warnln!(
+                "No source files specified in a sourcegroup in manifest for {}.",
+                package_name
+            );
+        }
+        self.extra.iter().for_each(|(k, _)| {
+            warnln!(
+                "Ignoring unknown field `{}` in sources in manifest for {}.",
+                k,
+                package_name
+            );
+        });
         Ok(Sources {
             target: self.target.unwrap_or(TargetSpec::Wildcard),
             include_dirs: include_dirs?,
             defines,
-            files: files?,
+            files,
         })
     }
 }
@@ -690,7 +745,7 @@ impl<'de> Deserialize<'de> for PartialSourceFile {
 impl Validate for PartialSourceFile {
     type Output = SourceFile;
     type Error = Error;
-    fn validate(self) -> Result<SourceFile> {
+    fn validate(self, package_name: &str) -> Result<SourceFile> {
         match self {
             PartialSourceFile::File(path) => {
                 let env_path_buf = env_path_from_string(path.clone())?;
@@ -701,7 +756,9 @@ impl Validate for PartialSourceFile {
                     Err(Error::new(format!("Error: file {} doesn't exist", &path)))
                 }
             }
-            PartialSourceFile::Group(srcs) => Ok(SourceFile::Group(Box::new(srcs.validate()?))),
+            PartialSourceFile::Group(srcs) => {
+                Ok(SourceFile::Group(Box::new(srcs.validate(package_name)?)))
+            }
         }
     }
 }
@@ -765,6 +822,9 @@ pub struct PartialWorkspace {
     pub checkout_dir: Option<String>,
     /// The locally linked packages.
     pub package_links: Option<IndexMap<String, String>>,
+    /// Unknown extra fields
+    #[serde(flatten)]
+    extra: HashMap<String, Value>,
 }
 
 impl PrefixPaths for PartialWorkspace {
@@ -780,6 +840,7 @@ impl PrefixPaths for PartialWorkspace {
                 ),
                 None => None,
             },
+            extra: self.extra,
         })
     }
 }
@@ -787,13 +848,20 @@ impl PrefixPaths for PartialWorkspace {
 impl Validate for PartialWorkspace {
     type Output = Workspace;
     type Error = Error;
-    fn validate(self) -> Result<Workspace> {
+    fn validate(self, package_name: &str) -> Result<Workspace> {
         let package_links: Result<IndexMap<_, _>> = self
             .package_links
             .unwrap_or_default()
             .iter()
             .map(|(k, v)| Ok((env_path_from_string(k.to_string())?, v.clone())))
             .collect();
+        self.extra.iter().for_each(|(k, _)| {
+            warnln!(
+                "Ignoring unknown field `{}` in workspace configuration in manifest for {}.",
+                k,
+                package_name
+            );
+        });
         Ok(Workspace {
             checkout_dir: match self.checkout_dir {
                 Some(dir) => Some(env_path_from_string(dir)?),
@@ -951,7 +1019,7 @@ impl Merge for PartialConfig {
 impl Validate for PartialConfig {
     type Output = Config;
     type Error = Error;
-    fn validate(self) -> Result<Config> {
+    fn validate(self, package_name: &str) -> Result<Config> {
         Ok(Config {
             database: match self.database {
                 Some(db) => env_path_from_string(db)?,
@@ -962,14 +1030,14 @@ impl Validate for PartialConfig {
                 None => return Err(Error::new("Git command or path to binary not configured")),
             },
             overrides: match self.overrides {
-                Some(d) => d.validate().map_err(|(key, cause)| {
+                Some(d) => d.validate(package_name).map_err(|(key, cause)| {
                     Error::chain(format!("In override `{}`:", key), cause)
                 })?,
                 None => IndexMap::new(),
             },
             plugins: match self.plugins {
                 Some(d) => d
-                    .validate()
+                    .validate(package_name)
                     .map_err(|(key, cause)| Error::chain(format!("In plugin `{}`:", key), cause))?,
                 None => IndexMap::new(),
             },
@@ -1090,7 +1158,7 @@ impl PrefixPaths for PartialVendorPackage {
 impl Validate for PartialVendorPackage {
     type Output = VendorPackage;
     type Error = Error;
-    fn validate(self) -> Result<VendorPackage> {
+    fn validate(self, package_name: &str) -> Result<VendorPackage> {
         Ok(VendorPackage {
             name: match self.name {
                 Some(name) => name,
@@ -1101,7 +1169,7 @@ impl Validate for PartialVendorPackage {
                 None => return Err(Error::new("external import target dir missing")),
             },
             upstream: match self.upstream {
-                Some(upstream) => upstream.validate().map_err(|cause| {
+                Some(upstream) => upstream.validate(package_name).map_err(|cause| {
                     Error::chain("Unable to parse external import upstream", cause)
                 })?,
                 None => return Err(Error::new("external import upstream missing")),