+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Add `clean` command to clean all bender related dependencies.
 - `script`: Allow flist formats to use `only-sources`/`only-includes`/`only-defines` flags.
 - Add check to ensure files referenced in all manifests exist.
+- Add warnings for unknown fields in manifest.
 
 ### Changed
 - `update`: Clean up alignment of manual resolution output.